simplify dependency between runtime and syntactic structure
